Fixed assertion when creating new perspective.
Wx was attempting to format a string with specifier %d (integer) with a
size_t.

+++ Source/Core/DolphinWX/FrameAui.cpp
@@ -533,8 +533,7 @@ void CFrame::OnPerspectiveMenu(wxCommandEvent& event)
 				wxTextEntryDialog dlg(this,
 						_("Enter a name for the new perspective:"),
 						_("Create new perspective"));
-				wxString DefaultValue = wxString::Format(_("Perspective %d"),
-						Perspectives.size() + 1);
+				wxString DefaultValue = wxString::Format(_("Perspective %d"), (int)(Perspectives.size() + 1));
 				dlg.SetValue(DefaultValue);
 
 				int Return = 0;
